Refactor the content replayer as a class
this allows to make it actually benefit from multithreading, by not
sharing objstorage instances between threads.

Note: the doctest formely in process_replay_objects_content's docstring
has been moved to a proper test in test_replay.py; the required mocking
to make it work is not suitable for docstring anymore.